create linear module
